build_image: btrfs: minimize allocation
Rebalance the /usr btrfs allocation to the maximum possible, in order to
increase the chance of having the btrfs `Free (statfs, df)` similar to
`Free (estimated)`.

Note that /usr is also a zstd compressed btrfs partition, so the output
of `df` free size and the actual free size after a file write for
example, will be very different, because the data in that file write has
a compression rate only definable after the file sync.

Unfortunately, there is no determinism in the btrfs file system case, because even if
you could in theory pre-compress with zstd the file before, and have an
idea about the size to be used, you still cannot really predict also the metadata
size for that file write.

See: https://github.com/flatcar/Flatcar/issues/1473

Signed-off-by: Adrian Vladu <avladu@cloudbasesolutions.com>