flatcar_workon: Bring back the --host flag
Telling flatcar_workon to work on SDK instead of a board could be
implemented either by passing an empty value for --board flag and
modifying `${FLAGS_board+foo}` to `${FLAGS_board:+foo}` or by adding
the `--host` boolean flag and killing the value of `--board` flag when
`--host` is true.

I went with the latter as it clearly conveys meaning at the call
sites.